# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API Backend NeoPay - Sistem Antrian Digital Bank
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /antrian/status-saya:
    get:
      operationId: getStatusSaya
      tags: [antrian]
      summary: Status antrian saya
      description: Melihat posisi dan status antrian milik nasabah yang login
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Status antrian nasabah
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusAntrianResponse"
        "404":
          description: Tidak ada antrian aktif
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /antrian/ambil:
    post:
      operationId: ambilAntrian
      tags: [antrian]
      summary: Ambil nomor antrian
      description: Nasabah mengambil nomor antrian hari ini
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AmbilAntrianBody"
      responses:
        "201":
          description: Berhasil ambil antrian
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AntrianDetailResponse"
        "400":
          description: Bank tutup atau kuota penuh
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Sudah punya antrian aktif hari ini
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /antrian/{id}/batal:
    put:
      operationId: batalAntrian
      tags: [antrian]
      summary: Batalkan antrian
      description: Nasabah membatalkan antriannya sendiri
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Antrian berhasil dibatalkan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AntrianDetailResponse"
        "404":
          description: Antrian tidak ditemukan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        pesan:
          type: string
        error:
          type: string
      required:
        - pesan
...
    AntrianPublikItem:
      type: object
      description: Data antrian publik tanpa PII (untuk tampilan layar tunggu)
      properties:
        id:
          type: string
        tanggal:
          type: string
        nomorAntrian:
          type: integer
        jenisLayanan:
          type: string
          enum: [teller, cs, kredit]
        status:
          type: string
          enum: [menunggu, dipanggil, sedang_dilayani, selesai, dibatalkan]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - tanggal
        - nomorAntrian
        - jenisLayanan
        - status
        - createdAt
        - updatedAt
    AntrianItem:
      type: object
      description: Data antrian lengkap dengan nasabahId (untuk nasabah atau admin yang terautentikasi)
      properties:
        id:
          type: string
        nasabahId:
          type: string
        cabangId:
          type: string
          nullable: true
        tanggal:
          type: string
        nomorAntrian:
          type: integer
        jenisLayanan:
          type: string
          enum: [teller, cs, kredit]
        status:
          type: string
          enum: [menunggu, dipanggil, sedang_dilayani, selesai, dibatalkan]
        catatan:
          type: string
          nullable: true
        dilayaniOlehId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        namaNasabah:
          type: string
          nullable: true
        noHpNasabah:
          type: string
          nullable: true
        namaCabang:
          type: string
          nullable: true
        alamatCabang:
          type: string
          nullable: true
        kodeCabang:
          type: string
          nullable: true
        namaPetugas:
          type: string
          nullable: true
      required:
        - id
        - nasabahId
        - tanggal
        - nomorAntrian
        - jenisLayanan
        - status
        - createdAt
        - updatedAt
    AntrianDetailResponse:
      type: object
      properties:
        antrian:
          $ref: "#/components/schemas/AntrianItem"
        posisi:
          type: integer
          nullable: true
      required:
        - antrian
    StatusAntrianResponse:
      type: object
      properties:
        antrian:
          $ref: "#/components/schemas/AntrianItem"
        posisi:
          type: integer
        totalMenunggu:
          type: integer
        dipanggilSekarang:
          allOf:
            - $ref: "#/components/schemas/AntrianPublikItem"
          nullable: true
      required:
        - antrian
        - posisi
        - totalMenunggu
...
    AmbilAntrianBody:
      type: object
      properties:
        jenisLayanan:
          type: string
          enum: [teller, cs, kredit]
        catatan:
          type: string
...