# lib/api-zod/src/generated/types/statusAntrianResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API Backend NeoPay - Sistem Antrian Digital Bank
 * OpenAPI spec version: 0.1.0
 */
import type { AntrianItem } from "./antrianItem";
import type { AntrianPublikItem } from "./antrianPublikItem";

export interface StatusAntrianResponse {
  antrian: AntrianItem;
  posisi: number;
  totalMenunggu: number;
  dipanggilSekarang?: AntrianPublikItem | null;
}